added new file to house warnings logic
