+++ src/components/utilities/index.js
@@ -1,6 +1,7 @@
 export { default as getLicenseAmendments } from './getLicenseAmendments';
 export { default as getResourceFromEntitlement } from './getResourceFromEntitlement';
 export { default as getResourceIdentifier } from './getResourceIdentifier';
+export { default as getConflictWarnings } from './getConflictWarnings';
 export { default as formatNoteReferrerEntityData } from './formatNoteReferrerEntityData';
 export { default as isExternal } from './isExternal';
 export { default as isPackage } from './isPackage';
