+++ src/components/utilities/validators.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { get } from 'lodash';
 import { FormattedMessage } from 'react-intl';
-import statuses from '../../constants/statuses';
 
 const required = value => (
   !value ? <FormattedMessage id="stripes-core.label.missingRequiredField" /> : undefined
@@ -109,24 +108,10 @@ const overlappingDates = (value, allValues, meta, errorMessageKey) => {
   return undefined;
 };
 
-const amendmentWarning = (statusInAgreement, statusInLicense, statusInLicenseLabel, startDate, endDate) => {
-  if (statusInAgreement === statuses.CURRENT) {
-    if (statusInLicense === statuses.EXPIRED || statusInLicense === statuses.REJECTED) {
-      return <FormattedMessage id="ui-agreements.license.warn.amendmentStatus" values={{ status: statusInLicenseLabel }} />;
-    } else if (startDate && new Date(startDate).getTime() > new Date().getTime()) {
-      return <FormattedMessage id="ui-agreements.license.warn.amendmentFuture" />;
-    } else if (endDate && new Date(endDate).getTime() < new Date().getTime()) {
-      return <FormattedMessage id="ui-agreements.license.warn.amendmentPast" />;
-    }
-  }
-  return null;
-};
-
 export default {
   dateOrder,
   multipleOpenEnded,
   overlappingDates,
   required,
   requiredStartDate,
-  amendmentWarning,
 };
